NIFI-11294 Extracted Ownership and Checkpoint conversion methods to a separate utility class
